### 2021-05-10 #3271 @whedon
[PRE REVIEW]: academictwitteR: an R Package to Access the Twitter Academic Research Product Track V2 API Endpoint
**Submitting author:** @cjbarrie (<a href="http://orcid.org/0000-0002-9156-990X">Christopher Barrie</a>)
**Repository:** <a href="https://github.com/cjbarrie/academictwitteR" target ="_blank">https://github.com/cjbarrie/academictwitteR</a>
**Version:** v0.1.0
