### 2021-05-10 #3271 @medewitt
Happy to review as well.

On Mon, May 10, 2021, 5:06 PM Daniel S. Katz ***@***.***>
wrote:

> We try for a couple of weeks, but a bit longer is ok. Reviews are somewhat
> iterative as well, so if there are any issues you find, it's great for you
> to point them out then rather than waiting to have a complete review. Is
> this ok? If so, I will add you as a reviewer, but the review won't start
> until I get another reviewer as well, and open a new issue for the review
> itself.
>
