+++ man/get_user_following.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_user_follows.R
+\name{get_user_following}
+\alias{get_user_following}
+\title{Get user following}
+\usage{
+get_user_following(x, bearer_token, ...)
+}
+\arguments{
+\item{x}{string containing one user id or a vector of user ids}
+
+\item{bearer_token}{string, bearer token}
+
+\item{...}{arguments passed to other backend functions}
+}
+\value{
+a data frame
+}
+\description{
+This function fetches a list of users the specified user ID is following.
+}
+\examples{
+\dontrun{
+bearer_token <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
+users <- "2244994945"
+get_user_following(users, bearer_token)
+}
+}
