+++ man/get_user_following.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_user_follows.R
+% Please edit documentation in R/get_user_edges.R
 \name{get_user_following}
 \alias{get_user_following}
 \title{Get user following}
