+++ man/get_all_tweets.Rd
@@ -13,6 +13,7 @@ get_all_tweets(
   data_path = NULL,
   export_query = TRUE,
   bind_tweets = TRUE,
+  page_n = 500,
   verbose = TRUE,
   ...
 )
@@ -33,6 +34,8 @@ get_all_tweets(
 \item{export_query}{If \code{TRUE}, query are exported to data_path}
 
 \item{bind_tweets}{If \code{TRUE}, tweets captured are bound into a data.frame for assignment}
+
+\item{page_n}{integer, amount of tweets to be returned by per page}
 
 \item{verbose}{If \code{FALSE}, query progress messages are suppressed}
 
