+++ man/get_liked_tweets.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_liked_tweets.R
+\name{get_liked_tweets}
+\alias{get_liked_tweets}
+\title{Get liked tweets}
+\usage{
+get_liked_tweets(x, bearer_token, ...)
+}
+\arguments{
+\item{x}{string containing one user id or a vector of user ids}
+
+\item{bearer_token}{string, bearer token}
+
+\item{...}{arguments passed to other backend functions}
+}
+\value{
+a data frame
+}
+\description{
+This function fetches a list of tweets liked by a user or users.
+}
+\examples{
+\dontrun{
+bearer_token <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
+users <- "2244994945"
+get_liked_tweets(users, bearer_token)
+}
+}
