+++ man/get_liking_users.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_liking_users.R
+\name{get_liking_users}
+\alias{get_liking_users}
+\title{Get liking users}
+\usage{
+get_liking_users(x, bearer_token, verbose = TRUE)
+}
+\arguments{
+\item{x}{string containing one tweet id or a vector of tweet ids}
+
+\item{bearer_token}{string, bearer token}
+
+\item{verbose}{If \code{FALSE}, query progress messages are suppressed}
+}
+\value{
+a data frame
+}
+\description{
+This function fetches a list of users who liked a tweet or tweets.
+}
+\examples{
+\dontrun{
+bearer_token <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
+tweet <- "1387744422729748486"
+get_liking_users(tweet, bearer_token)
+}
+}
