+++ NAMESPACE
@@ -9,6 +9,8 @@ export(get_bbox_tweets)
 export(get_country_tweets)
 export(get_geo_tweets)
 export(get_lang_tweets)
+export(get_liked_tweets)
+export(get_liking_users)
 export(get_media_tweets)
 export(get_mentions_tweets)
 export(get_place_tweets)
