### 2021-05-24 97bc482
allowing github actions and examples of testing framework
