### 2021-05-24 cjbarrie/academictwitteR#78 @medewitt
Automated Testing / Unit Testing Framework for JOSS
CRAN build checks appear to be the only automated testing implemented in the package. Additionally, many of the examples in the documentation are wrapped in "don't run" which means they are not checked on CRAN (which makes sense for CRAN checks, but not for ensuring unit tests).

**Describe the solution you'd like**
>Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Ideally, continuous integration could be deployed to verify continuity of functionality with new pulls. This could be implemented through GitHub Actions. Additionally, automated tests could be written to verify that functions are working properly. This is challenging with API related packages, but it could be done through minimally testing non-API key required features or with the use of secrets, some basic api features could be tested. And example is available in a forthcoming PR. Totally open to pushback on this point.

Testthat or tinytest both offer good options for unit testing.

**Additional context**
See <https://github.com/openjournals/joss-reviews/issues/3272>
