### 2021-05-24 cjbarrie/academictwitteR#78 @medewitt
See https://github.com/cjbarrie/academictwitteR/pull/79#issue-651480789 for an example of continuous integration and unit testing.
