### 2021-05-24 cjbarrie/academictwitteR#80 @medewitt
Contributor Code of Conduct
>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I don't see a contributor code of conduct on the README page or in the vignettes.

**Describe the solution you'd like**

Clear guidelines for those wishing to contribute to the package.

The `usethis` package has a pretty vanilla contributor code of conduct that represents a good starting place.
You can test it out using the [`usethis` package](https://usethis.r-lib.org/reference/use_code_of_conduct.html)

Checkout:

```r
usethis::use_code_of_conduct()
```

**Additional context**

Part of JOSS review see https://github.com/openjournals/joss-reviews/issues/3272
