+++ man/get_all_tweets.Rd
@@ -9,6 +9,7 @@ get_all_tweets(
   start_tweets,
   end_tweets,
   bearer_token,
+  n = 100,
   file = NULL,
   data_path = NULL,
   export_query = TRUE,
@@ -26,6 +27,8 @@ get_all_tweets(
 \item{end_tweets}{string, ending date}
 
 \item{bearer_token}{string, bearer token}
+
+\item{n}{integer, amount of tweets to be fetched}
 
 \item{file}{string, name of the resulting RDS file}
 
