### 2021-05-24 #3272 @JosiahParry
@danielskatz quick point of clarification (and point me to any documentation, please): must the `LICENSE` contain the full content of the license? Or is the LICENSE.md containing that information sufficient?
