### 2021-05-24 #3272 @danielskatz
I'm not sure it matters if the file is `LICENSE` or `LICENSE.md` but having both with different contents seems bad to me in my Python world.  I would say that `LICENSE` in this case should be removed, but CRAN has policies that seem odd to me, specifically see the MIT link in https://cran.r-project.org/web/licenses/
