### 2021-05-25 #3272 @JosiahParry
@cjbarrie, you're missing a code of conduct. I'd recommend starting with a template from something like `usethis::use_code_of_conduct()`
