### 2021-05-25 #3272 @medewitt
@danielskatz still working through the review on my end. I have put in two PRs

[Update to Cite the JOSS Article for rtweet](https://github.com/cjbarrie/academictwitteR/pull/82)
[Example of Unit Tests/ Continuous Integration Through GH Actions](https://github.com/cjbarrie/academictwitteR/pull/79)

Additionally, I have an issue on the repository for with recommendation to use the code of conduct:

[Establishing Code of Conduct/Contribution Guides](https://github.com/cjbarrie/academictwitteR/issues/80)
