### 2021-05-25 cjbarrie/academictwitteR#83 @medewitt
Continuous Integration Fails on R Versions < 3.4
This I learned, `isFALSE` was introduced in R3.4.2. Right now in your DESCRIPTION file, you have R > 3 as the only restriction on the version of R users are required to have. In order to ensure this works for earlier versions of R you would need to change the following `isFALSE` sections to `!isTRUE` or more robust`(!is.null(x) && !isTRUE(x))`

https://github.com/cjbarrie/academictwitteR/blob/3a2d429f428b2dea359f41b5d7b3a50f315b25de/R/build_user_queryv2.R#L76

https://github.com/cjbarrie/academictwitteR/blob/3a2d429f428b2dea359f41b5d7b3a50f315b25de/R/build_queryv2.R#L85

The CRAN checks missed this because they only check 1 version past, current version, and devel versions of R.

https://github.com/openjournals/joss-reviews/issues/3272
