### 2021-05-25 9c7bb09
Reworked get_users_tweets() and related functions. Added add_query_prefix() to handle query vector
