### 2021-05-25 7ff87a9
Fix isFALSE issue for R > 3.4.2
