+++ R/build_user_queryv2.R
@@ -1,200 +1,0 @@
-#' Build tweet query 
-#' 
-#' Build tweet query according to targeted parameters, can then be input to main \code{\link{get_user_tweets}} function as query parameter.
-#'
-#' @param users string or character vector, user handles from which to collect data
-#' @param is_retweet If `TRUE`, only retweets will be returned; if `FALSE` retweets will not be returned
-#' @param is_reply If `TRUE`, only reply tweets will be returned
-#' @param is_quote If `TRUE`, only quote tweets will be returned
-#' @param place string, name of place e.g. "London"
-#' @param country string, name of country as ISO alpha-2 code e.g. "GB"
-#' @param point_radius numeric, a vector of two point coordinates latitude, longitude, and point radius distance (in miles)
-#' @param bbox numeric, a vector of four bounding box coordinates from west longitude to north latitude
-#' @param geo_query If `TRUE` user will be prompted to enter relevant information for bounding box or point radius geo buffers
-#' @param remove_promoted If `TRUE`, tweets created for promotion only on ads.twitter.com are removed
-#' @param has_hashtags If `TRUE`, only tweets containing hashtags will be returned
-#' @param has_cashtags If `TRUE`, only tweets containing cashtags will be returned
-#' @param has_links If `TRUE`, only tweets containing links and media will be returned
-#' @param has_mentions If `TRUE`, only tweets containing mentions will be returned
-#' @param has_media If `TRUE`, only tweets containing a recognized media object, such as a photo, GIF, or video, as determined by Twitter will be returned
-#' @param has_images If `TRUE`, only tweets containing a recognized URL to an image will be returned
-#' @param has_videos If `TRUE`, only tweets containing contain native Twitter videos, uploaded directly to Twitter will be returned
-#' @param has_geo If `TRUE`, only tweets containing Tweet-specific geolocation data provided by the Twitter user will be returned
-#' @param lang string, a single BCP 47 language identifier e.g. "fr"
-#'
-#' @return a query string
-#' @export
-#'
-#' @examples
-#' \dontrun{
-#' users <- c("cbarrie", "justin_ct_ho")
-#' query <-   build_user_query(users, is_retweet = F, has_media = T, lang = "en")
-#' }
-#' 
-#' @importFrom utils menu
-#' 
-
-build_user_query <- function(users, 
-                        is_retweet = NULL, 
-                        is_reply = FALSE, 
-                        is_quote = FALSE,
-                        place = NULL, 
-                        country = NULL, 
-                        point_radius = NULL,
-                        bbox = NULL,
-                        geo_query = FALSE,
-                        remove_promoted = FALSE,
-                        has_hashtags = FALSE,
-                        has_cashtags = FALSE,
-                        has_links = FALSE,
-                        has_mentions = FALSE,
-                        has_media = FALSE,
-                        has_images = FALSE,
-                        has_videos = FALSE,
-                        has_geo = FALSE,
-                        lang= NULL) {
-  
-  for(i in seq_along(users)) {
-    query <- users[[i]]
-  
-  if (isTRUE(is_retweet) & isTRUE(is_reply)) {
-    stop("A tweet cannot be both a retweet and a reply")
-  }
-  
[…]
