### 2021-05-25 cjbarrie/academictwitteR#83 @justinchuntingho
Thanks a lot! This is now fixed with #85
We changed the `isFALSE` sections to `!isTRUE` in `build_query()` and retired `build_user_query()` as its funtionalies are now replaced by other functions.
