+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: academictwitteR
 Title: Access the Twitter Academic Research Product Track V2 API Endpoint
-Version: 0.1.0
+Version: 0.2.0
 Authors@R: 
     c(person(given = "Christopher",
            family = "Barrie",
