### 2021-05-25 cjbarrie/academictwitteR#80 @justinchuntingho
Thanks for pointing this out! Added Code of Conduct in PR #86
