### 2021-05-25 #3272 @karthik
Yes CRAN does not like the markdown version. If you include it, you should add it to `.Rbuildignore` so it is not bundled with the built package. For the most up to date information on LICENSE for R packages, see https://r-pkgs.org/license.html#key-files

If you are not sure how to set these up, I would recommend using the `usethis` package and its relevant functions: https://r-pkgs.org/license.html#more-licenses
