### 2021-05-26 cjbarrie/academictwitteR#78 @justinchuntingho
Thanks for pointing out. Merged #79 for the unit testing.
