### 2021-05-28 #3272 @JosiahParry
It looks like the code of conduct has been added.
The license looks set up accordingly (thanks, @karthik).

Will review further on Monday.
