### 2021-06-02 #3272 @danielskatz
The checkbox item says "Are there automated tests or manual steps described so that the functionality of the software can be verified?"  So we don't need automated tests, if there is a way to verify the functionality of the software manually.  Is there such a way documented?
