### 2021-06-02 #3272 @medewitt
Check out https://github.com/cjbarrie/academictwitteR/issues/78#issue-899836077

Right now, the CRAN checks will do the basic installation checks, but will not check the functionality per se (all of the examples which would be checked by CRAN are wrapped in "dontrun").  As far as manually checking functionality, the authors have done a great job adding an example to each function. A user could manually try to run each example to demonstrate functionality.
