### 2021-06-02 #3272 @medewitt
The testing piece is it on my end. If you're ok with the described manual checking and the existing CI that runs the cran checks across platforms, I can check the final box.
