### 2021-06-02 #3272 @danielskatz
Yes, I think what you described meets the description/criterion
