### 2021-06-03 #3272 @cjbarrie
> The testing piece is it on my end. If you're ok with the described manual checking and the existing CI that runs the cran checks across platforms, I can check the final box.

Thanks @medewitt. Have merged your PR for testing JSON bind too, and including more informative error messaging. Very grateful for your help with this!
