### 2021-06-04 69e4c3c
Add set_bearer and get_bearer
