+++ R/get_lang_tweetsv2.R
@@ -32,7 +32,7 @@ get_lang_tweets <-
            lang,
            start_tweets,
            end_tweets,
-           bearer_token,
+           bearer_token = get_bearer(),
            file = NULL,
            data_path = NULL,
            bind_tweets = TRUE,
