+++ R/get_geo_tweetsv2.R
@@ -30,7 +30,7 @@ get_geo_tweets <-
   function(query,
            start_tweets,
            end_tweets,
-           bearer_token,
+           bearer_token = get_bearer(),
            file = NULL,
            data_path = NULL,
            bind_tweets = TRUE,
