+++ R/get_retweets_of_userv2.R
@@ -30,7 +30,7 @@ get_retweets_of_user <-
   function(users,
            start_tweets,
            end_tweets,
-           bearer_token,
+           bearer_token = get_bearer(),
            n = 100,
            file = NULL,
            data_path = NULL,
