+++ R/get_bbox_tweetsv2.R
@@ -33,7 +33,7 @@ get_bbox_tweets <-
            bbox,
            start_tweets,
            end_tweets,
-           bearer_token,
+           bearer_token = get_bearer(),
            file = NULL,
            data_path = NULL,
            bind_tweets = TRUE,
