+++ man/get_all_tweets.Rd
@@ -8,7 +8,7 @@ get_all_tweets(
   query = NULL,
   start_tweets,
   end_tweets,
-  bearer_token,
+  bearer_token = get_bearer(),
   n = 100,
   file = NULL,
   data_path = NULL,
