### 2021-06-06 7a16b18
Don't delete aka unlink the session's temp directory; clean up after testing
