+++ tests/testthat/test-bind_tweet_jsons.R
@@ -1,6 +1,7 @@
 ## context("bind_tweet_jsons")
 
-empty_dir <- tempdir()
+empty_dir <- paste0(tempdir(), "/", paste0(sample(letters, 20), collapse = ""))
+dir.create(empty_dir)
 
 my_cars <- mtcars
 my_cars$model <- rownames(my_cars)
@@ -16,7 +17,9 @@ test_that("Expect sucess in binding two jsons", {
 })
 
 unlink(empty_dir, recursive = TRUE)
-temp_dir <- tempdir()
+
+temp_dir <-  paste0(tempdir(), "/", paste0(sample(letters, 20), collapse = ""))
+dir.create(temp_dir)
 
 test_that("Error on finding no jsons to bind", {
   expect_error(bind_tweet_jsons(temp_dir))
@@ -30,3 +33,5 @@ test_that("real data", {
 test_that("trailing slash", {
   expect_error(bind_tweet_jsons("../testdata/commtwitter/"), NA)
 })
+
+unlink(temp_dir)
