### 2021-06-07 #3272 @cjbarrie
Thank you very much for this!

- Tag: v0.1.1.1 released as "JOSS Release" here: https://github.com/cjbarrie/academictwitteR/releases/tag/v0.1.1.1
- Archived at Zenodo here: https://zenodo.org/record/4905468
- DOI of archived version: https://doi.org/10.5281/zenodo.4905468
