### 2021-06-07 5de006d
Make GH action doesn't check for ancient versions of R
