### 2021-06-07 b31b883
clarify the version requirement and add GH workflow
