+++ .github/workflows/R-CMD-check.yaml
@@ -24,6 +24,8 @@ jobs:
           - {os: macOS-latest, r: 'release'}
           - {os: ubuntu-20.04, r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
           - {os: ubuntu-20.04, r: 'devel', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
+          - {os: ubuntu-20.04, r: 'oldrel', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
+          - {os: ubuntu-18.04, r: '3.4', "https://packagemanager.rstudio.com/cran/__linux__/bionic/latest"}
           
     env:
       RSPM: ${{ matrix.config.rspm }}
