+++ DESCRIPTION
@@ -30,6 +30,8 @@ BugReports: https://github.com/cjbarrie/academictwitteR/issues
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.1.1
+Depends:
+    R (>= 3.4)
 Imports: 
     dplyr,
     httr,
