### 2021-06-07 #3272 @cjbarrie
> @cjbarrie - Is this ok for you? Or did you want this to be "JOSS Release", which seems a bit unusual to me?

This is fine, thank you! Apologies for the confusion.
