### 2021-06-07 #3272 @cjbarrie
Fantastic! Thanks so much for all your help with this. I will be letting colleagues know of JOSS. Honestly so impressed with this review process.

What a brilliant initiative. I will, of course, be adding my name to the list of potential reviewers.

Chris

> On 7 Jun 2021, at 20:17, whedon ***@***.***> wrote:
>
> ﻿
> 🎉🎉🎉 Congratulations on your paper acceptance! 🎉🎉🎉
>
> If you would like to include a link to your paper from your README use the following code snippets:
>
> Markdown:
>
> HTML:
> <a style="border-width:0" href="https://doi.org/10.21105/joss.03272">
> </a>
>
> reStructuredText:
>    :target: https://doi.org/10.21105/joss.03272
> This is how it will look in your documentation:
>
>
>
> We need your help!
>
> Journal of Open Source Software is a community-run journal and relies upon volunteer effort. If you'd like to support us please consider doing either one (or both) of the the following:
>
> Volunteering to review for us sometime in the future. You can add your name to the reviewer list here: https://joss.theoj.org/reviewer-signup.html
> Making a small donation to support our running costs here: https://numfocus.org/donate-to-joss
